correct legacy instance-initializer support
