+++ app/initializers/head-tags.js
@@ -1,1 +1,20 @@
-export { default, initialize } from 'ember-cli-meta-tags/initializers/head-tags';
+import instanceInitializer from '../instance-initializers/head-tags';
+
+export function initialize(registry, application) {
+  application.inject('service:head-tags', 'router', 'router:main');
+
+  // Ember >= 1.12
+  if (application.instanceInitializer) {
+    return;
+  }
+
+  // Ember < 1.12
+  instanceInitializer.initialize({
+    container: registry
+  });
+}
+
+export default {
+  name: 'head-tags',
+  initialize: initialize
+};
